Squashed commit of the following:

commit 0901ff06f4cb519563c6d5a3aaba6e23b05e7c1c
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Tue Dec 15 10:05:01 2020 -0800

    address PR comments

commit 5e1860678b98d4710efa13302b7b1855ee701cdb
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Tue Dec 15 03:18:10 2020 -0500

    move tar log file

commit 9f5e866d9daee8b84da34ad17f28f53911615e75
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Tue Dec 15 01:17:22 2020 -0500

    update deprecated functions

commit e5fd8798e4a264c4772f9d660c6436348aa4c878
Author: David J. Gardner <gardner48@llnl.gov>
Date:   Tue Dec 15 00:55:04 2020 -0500

    array len must be larger than string length for null terminator

+++ examples/nvector/cuda/test_nvector_cuda.cu
@@ -367,7 +367,8 @@ int check_ans(realtype ans, N_Vector X, sunindextype length)
   /* check vector data */
   for (i = 0; i < length; i++) {
     if (failure += FNEQ(Xdata[i], ans)) {
-      printf("check_ans fail: Xdata[%d] = %f, expected Xdata[%d] = %f\n", i, Xdata[i], i, ans);
+      printf("check_ans fail: Xdata[%ld] = %f, expected Xdata[%ld] = %f\n",
+             (long int)i, Xdata[i], (long int)i, ans);
     }
   }
